fix: correct connected_account_id FK column type in bindings migration
The connected_account_id column was declared as unsignedInteger while
connected_accounts.id is a bigint. MariaDB rejects the foreign key
(errno 150) because integer FK columns must match the referenced type
exactly. Use unsignedBigInteger instead.

+++ database/migrations/2026_05_13_200001_create_task_provider_bindings_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreign('repo_profile_id')->references('id')->on('repo_profiles')->cascadeOnDelete();
             $table->string('kind');
             $table->string('mode')->default('disabled');
-            $table->unsignedInteger('connected_account_id')->nullable();
+            $table->unsignedBigInteger('connected_account_id')->nullable();
             $table->foreign('connected_account_id')->references('id')->on('connected_accounts')->nullOnDelete();
             $table->string('external_project_ref')->nullable();
             $table->json('filters')->nullable();
